Add home-upcoming.js
