+++ src/js/index.js
@@ -25,7 +25,7 @@ import "./other/current-page-track"
 
 import "./other/scroll-up"
 import "./other/weekly-trends.js"
-
+import "./other/home-upcoming"
 
 import "./other/test-card-markup"
 import "./other/switch-theme-btn"
